refactor: корректировка тестов с учетом новых изменений.
