+++ src/test/java/ru/yandex/practicum/filmorate/controllerTest/FilmControllerTest.java
@@ -1,24 +1,26 @@
-package ru.yandex.practicum.filmorate;
+package ru.yandex.practicum.filmorate.controllerTest;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import ru.yandex.practicum.filmorate.controller.FilmController;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.service.FilmService;
 
 import java.time.LocalDate;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(FilmController.class)
 public class FilmControllerTest extends BaseControllerTest {
+    @MockBean
+    private FilmService filmService;
+
     public FilmControllerTest() {
         endPoint = "/films";
+
     }
 
     // Post /films
@@ -31,21 +33,13 @@ public class FilmControllerTest extends BaseControllerTest {
                 .duration(60)
                 .build();
 
-        assertEquals(0, film.getId());
 
         final String json = asJsonString(film);
 
-        final MvcResult result = this.mockMvc.perform(MockMvcRequestBuilders.post(endPoint)
+        this.mockMvc.perform(MockMvcRequestBuilders.post(endPoint)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(json))
-                .andExpect(status().isCreated())
-                .andReturn();
-
-        final String responseJson = result.getResponse().getContentAsString();
-        final Film responseFilm = objectMapper.readValue(responseJson, Film.class);
-
-        assertNotEquals(0, responseFilm.getId());
-        assertThat(film).usingRecursiveComparison().ignoringFields("id").isEqualTo(responseFilm);
+                .andExpect(status().isCreated());
     }
 
     // Put /users
@@ -58,8 +52,7 @@ public class FilmControllerTest extends BaseControllerTest {
                 .duration(60)
                 .build();
 
-        MvcResult result = addEntity(film);
-        film = objectMapper.readValue(result.getResponse().getContentAsString(), Film.class);
+        addEntity(film);
 
         film.setName("new name");
         film.setDescription("new description");
@@ -67,16 +60,9 @@ public class FilmControllerTest extends BaseControllerTest {
         film.setReleaseDate(LocalDate.now().minusDays(60));
 
         final String json = asJsonString(film);
-        result = this.mockMvc.perform(MockMvcRequestBuilders.put(endPoint)
+        this.mockMvc.perform(MockMvcRequestBuilders.put(endPoint)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(json))
-                .andExpect(status().isOk())
-                .andReturn();
-
-        final String responseJson = result.getResponse().getContentAsString();
-        final Film responseFilm = objectMapper.readValue(responseJson, Film.class);
-
-        assertEquals(film.getId(), responseFilm.getId());
-        assertThat(film).usingRecursiveComparison().ignoringFields("id").isEqualTo(responseFilm);
+                .andExpect(status().isOk());
     }
 }
